+++ src/Installation/InstallationCreatedAt.php
@@ -7,17 +7,17 @@ namespace DevboardLib\GitHub\Installation;
 use DateTime;
 
 /**
- * @see CreatedAtSpec
- * @see CreatedAtTest
+ * @see InstallationCreatedAtSpec
+ * @see InstallationCreatedAtTest
  */
-class CreatedAt extends DateTime
+class InstallationCreatedAt extends DateTime
 {
     public function __toString(): string
     {
         return $this->format('c');
     }
 
-    public static function createFromFormat($format, $time, $object = null): CreatedAt
+    public static function createFromFormat($format, $time, $object = null): InstallationCreatedAt
     {
         $date = parent::createFromFormat($format, $time, $object);