+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 1.0.0
 -----
 
+ * #113 Make Installation namespace conistent (prefix all with only `Installation`)
  *      Add pull request template
  * #111 Move `Application` related VO's to separate namespace (DevboardLib\GitHub\Installation\ApplicationId -> DevboardLib\GitHub\Application\ApplicationId)
  * #110 Moving from `Devboard` to `DevboardLib` namespace